Move dictionary-validate.js
